test: 테스트 코드 분리
- InputValidator 구현 중 구조 변경을 위해 수정
- 당첨 번호, 보너스 번호, 구매 금액으로 분리
- 각각에 대한 Validator 생성 예정
